+++ src/events/LinkClickEvent.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Masuga\LinkVault\events;
+
+use yii\base\Event;
+
+class LinkClickEvent extends Event
+{
+    /**
+     * The Link Vault parameters array.
+     */
+    public $parameters = [];
+}