+++ src/resources/ReportsAssetBundle.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Masuga\LinkVault\resources;
+
+use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
+
+class ReportsAssetBundle extends AssetBundle
+{
+	public function init()
+	{
+		$this->sourcePath = '@Masuga/LinkVault/resources';
+		$this->depends = [
+			CpAsset::class,
+		];
+		$this->css = [
+			'pagination.css',
+			'reports.css'
+		];
+		$this->js = [
+			'reports.js',
+		];
+		parent::init();
+	}
+}